+++ cardinal-components-base/src/main/java/org/ladysnake/cca/api/v3/component/Component.java
@@ -24,6 +24,7 @@ package org.ladysnake.cca.api.v3.component;
 
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.registry.RegistryWrapper;
 import org.jetbrains.annotations.Contract;
 
 /**
@@ -40,21 +41,23 @@ public interface Component {
     /**
      * Reads this component's properties from a {@link NbtCompound}.
      *
-     * @param tag a {@code NbtCompound} on which this component's serializable data has been written
+     * @param tag            a {@code NbtCompound} on which this component's serializable data has been written
+     * @param registryLookup access to dynamic registry data
      * @implNote implementations should not assert that the data written on the tag corresponds to any
      * specific scheme, as saved data is susceptible to external tempering, and may come from an earlier
      * version.
      */
     @Contract(mutates = "this")
-    void readFromNbt(NbtCompound tag);
+    void readFromNbt(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup);
 
     /**
      * Writes this component's properties to a {@link NbtCompound}.
      *
-     * @param tag a {@code NbtCompound} on which to write this component's serializable data
+     * @param tag            a {@code NbtCompound} on which to write this component's serializable data
+     * @param registryLookup access to dynamic registry data
      */
-    @Contract(mutates = "param")
-    void writeToNbt(NbtCompound tag);
+    @Contract(mutates = "param1")
+    void writeToNbt(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup);
 
     /**
      * Indicates whether some other object is "equal to" this component.