+++ cardinal-components-base/src/main/java/org/ladysnake/cca/api/v3/component/TransientComponent.java
@@ -23,6 +23,7 @@
 package org.ladysnake.cca.api.v3.component;
 
 import net.minecraft.nbt.NbtCompound;
+import net.minecraft.registry.RegistryWrapper;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -30,12 +31,12 @@ import org.jetbrains.annotations.Nullable;
  */
 public interface TransientComponent extends Component {
     @Override
-    default void readFromNbt(NbtCompound tag) {
+    default void readFromNbt(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup) {
         // Nothing to read
     }
 
     @Override
-    default void writeToNbt(NbtCompound tag) {
+    default void writeToNbt(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup) {
         // Nothing to write
     }