+++ cardinal-components-base/src/main/java/org/ladysnake/cca/internal/base/AbstractComponentContainer.java
@@ -25,6 +25,7 @@ package org.ladysnake.cca.internal.base;
 import net.fabricmc.fabric.api.util.NbtType;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtList;
+import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.util.Identifier;
 import org.ladysnake.cca.api.v3.component.Component;
 import org.ladysnake.cca.api.v3.component.ComponentContainer;
@@ -42,20 +43,20 @@ public abstract class AbstractComponentContainer implements ComponentContainer {
     public static final String NBT_KEY = "cardinal_components";
 
     @Override
-    public void copyFrom(ComponentContainer other) {
+    public void copyFrom(ComponentContainer other, RegistryWrapper.WrapperLookup registryLookup) {
         for (ComponentKey<?> key : this.keys()) {
             Component theirs = key.getInternal(other);
             Component ours = key.getInternal(this);
             assert ours != null;
 
             if (theirs != null && !ours.equals(theirs)) {
-                if (ours instanceof CopyableComponent) {
+                if (ours instanceof CopyableComponent<?>) {
                     @SuppressWarnings("unchecked") CopyableComponent<Component> copyable = (CopyableComponent<Component>) ours;
-                    copyable.copyFrom(theirs);
+                    copyable.copyFrom(theirs, registryLookup);
                 } else {
                     NbtCompound tag = new NbtCompound();
-                    theirs.writeToNbt(tag);
-                    ours.readFromNbt(tag);
+                    theirs.writeToNbt(tag, registryLookup);
+                    ours.readFromNbt(tag, registryLookup);
                 }
             }
         }
@@ -72,7 +73,7 @@ public abstract class AbstractComponentContainer implements ComponentContainer {
      * type, the component tag is skipped.
      */
     @Override
-    public void fromTag(NbtCompound tag) {
+    public void fromTag(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup) {
         if(tag.contains(NBT_KEY, NbtType.LIST)) {
             NbtList componentList = tag.getList(NBT_KEY, NbtType.COMPOUND);
             for (int i = 0; i < componentList.size(); i++) {
@@ -81,7 +82,7 @@ public abstract class AbstractComponentContainer implements ComponentContainer {
                 if (type != null) {
                     Component component = type.getInternal(this);
                     if (component != null) {
-                        component.readFromNbt(nbt);
+                        component.readFromNbt(nbt, registryLookup);
                     }
                 }
             }
@@ -94,7 +95,7 @@ public abstract class AbstractComponentContainer implements ComponentContainer {
                 if (componentMap.contains(keyId, NbtType.COMPOUND)) {
                     Component component = key.getInternal(this);
                     assert component != null;
-                    component.readFromNbt(componentMap.getCompound(keyId));
+                    component.readFromNbt(componentMap.getCompound(keyId), registryLookup);
                     componentMap.remove(keyId);
                 }
             }
@@ -109,19 +110,19 @@ public abstract class AbstractComponentContainer implements ComponentContainer {
      * @implSpec This implementation first checks if the container is empty; if so it
      * returns immediately. Then, it iterates over this container's mappings, and creates
      * a compound tag for each component. The tag is then passed to the component's
-     * {@link Component#writeToNbt(NbtCompound)} method. Every such serialized component is appended
+     * {@link Component#writeToNbt(NbtCompound, RegistryWrapper.WrapperLookup)} method. Every such serialized component is appended
      * to a {@code NbtCompound}, using the component type's identifier as the key.
      * The serialized map is finally appended to the passed in tag using the "cardinal_components" key.
      */
     @Override
-    public NbtCompound toTag(NbtCompound tag) {
+    public NbtCompound toTag(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup) {
         if(this.hasComponents()) {
             NbtCompound componentMap = null;
             NbtCompound componentTag = new NbtCompound();
 
             for (ComponentKey<?> type : this.keys()) {
                 Component component = type.getFromContainer(this);
-                component.writeToNbt(componentTag);
+                component.writeToNbt(componentTag, registryLookup);
 
                 if (!componentTag.isEmpty()) {
                     if (componentMap == null) {