+++ cardinal-components-entity/src/main/java/org/ladysnake/cca/mixin/entity/common/BucketableMixin.java
@@ -22,10 +22,8 @@
  */
 package org.ladysnake.cca.mixin.entity.common;
 
-import net.minecraft.component.DataComponentTypes;
 import net.minecraft.entity.Bucketable;
 import net.minecraft.entity.mob.MobEntity;
-import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NbtCompound;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -36,11 +34,11 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public interface BucketableMixin {
     @Inject(method = "method_57302", at = @At("RETURN"))
     private static void writeComponentsToStack(MobEntity mobEntity, NbtCompound nbtCompound, CallbackInfo ci) {
-        mobEntity.asComponentProvider().getComponentContainer().toTag(nbtCompound);
+        mobEntity.asComponentProvider().getComponentContainer().toTag(nbtCompound, mobEntity.getRegistryManager());
     }
 
     @Inject(method = "copyDataFromNbt(Lnet/minecraft/entity/mob/MobEntity;Lnet/minecraft/nbt/NbtCompound;)V", at = @At("RETURN"))
     private static void readComponentsFromStack(MobEntity entity, NbtCompound nbt, CallbackInfo ci) {
-        entity.asComponentProvider().getComponentContainer().fromTag(nbt);
+        entity.asComponentProvider().getComponentContainer().fromTag(nbt, entity.getRegistryManager());
     }
 }