+++ cardinal-components-scoreboard/src/main/java/org/ladysnake/cca/mixin/scoreboard/MixinScoreboardState.java
@@ -26,11 +26,9 @@ import com.llamalad7.mixinextras.sugar.Local;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtList;
 import net.minecraft.registry.RegistryWrapper;
-import net.minecraft.scoreboard.AbstractTeam;
 import net.minecraft.scoreboard.Scoreboard;
 import net.minecraft.scoreboard.ScoreboardState;
 import net.minecraft.scoreboard.Team;
-import net.minecraft.text.Text;
 import org.ladysnake.cca.api.v3.component.ComponentProvider;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -39,10 +37,6 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
-
-import java.util.Collection;
-import java.util.Iterator;
 
 @Mixin(ScoreboardState.class)
 public abstract class MixinScoreboardState {
@@ -51,13 +45,13 @@ public abstract class MixinScoreboardState {
     private Scoreboard scoreboard;
 
     @Inject(method = "writeNbt", at = @At("RETURN"))
-    private void saveComponents(NbtCompound tag, RegistryWrapper.WrapperLookup registryLookup, CallbackInfoReturnable<NbtCompound> cir) {
-        ((ComponentProvider) this.scoreboard).getComponentContainer().toTag(tag);
+    private void saveComponents(NbtCompound tag, RegistryWrapper.WrapperLookup registries, CallbackInfoReturnable<NbtCompound> cir) {
+        ((ComponentProvider) this.scoreboard).getComponentContainer().toTag(tag, registries);
     }
 
     @Inject(method = "readNbt", at = @At("RETURN"))
     private void loadComponents(NbtCompound tag, RegistryWrapper.WrapperLookup registries, CallbackInfoReturnable<ScoreboardState> cir) {
-        ((ComponentProvider) this.scoreboard).getComponentContainer().fromTag(tag);
+        ((ComponentProvider) this.scoreboard).getComponentContainer().fromTag(tag, registries);
     }
 
     @Inject(
@@ -69,7 +63,7 @@ public abstract class MixinScoreboardState {
         )
     )
     private void loadTeamComponents(NbtList nbt, RegistryWrapper.WrapperLookup registries, CallbackInfo ci, @Local NbtCompound teamData, @Local Team team) {
-        ((ComponentProvider) team).getComponentContainer().fromTag(teamData);
+        ((ComponentProvider) team).getComponentContainer().fromTag(teamData, registries);
     }
 
     @Inject(
@@ -80,10 +74,11 @@ public abstract class MixinScoreboardState {
         )
     )
     private void saveTeamComponents(
+        RegistryWrapper.WrapperLookup registries,
         CallbackInfoReturnable<NbtList> cir,
         @Local Team team,
         @Local NbtCompound teamData
     ) {
-        ((ComponentProvider) team).getComponentContainer().toTag(teamData);
+        ((ComponentProvider) team).getComponentContainer().toTag(teamData, registries);
     }
 }